Fix SC2154 of shellchecker (#498)
SC2154: var is referenced but not assigned.
See https://github.com/koalaman/shellcheck/wiki/SC2154

This part of the code does not seem finished.
Remove the sed's for the variables which don't exist, but keep the
original code as comment. When the relevant code is fixed the
commented code may be used again.

Signed-off-by: Matthias May <github@may.nu>