Hardcode JDK `downloadPath`
Azul changed their website and now it populates the list of JDKs
dynamically via Ajax call.

This PR hardcodes the 2 URLs in order to improve realiability of the
script and to point statically to a defined version so to avoid potential
compatibility issues. This also means that we will need to keep track
of new version and update the URLs.
We might consider extracting these URLs in two extranal constant defined
elsewhere.

Closes #642

Signed-off-by: Libo Cannici <lcannici@zendesk.com>