fix influx setup problems (#774)
* wait longer for influx to come up

Influxdb startup (even with an empty db or maybe because of it) can
take longer than 10 seconds. Wait for up to 30 seconds to be safe.

Signed-off-by: Andreas Schultz <aschultz@warp10.net>

* fix match in influx config sed script

Newer default configs are indented by two spaces. Remove the start
of line anchor to account for that.

Signed-off-by: Andreas Schultz <aschultz@warp10.net>

* preset influxdb_admin_username for local installations

The openhab DB installation tests for this, need to preset it for local
installations as well.

Signed-off-by: Andreas Schultz <aschultz@warp10.net>