Change java-jre.bash to use AdoptOpenJDK
After the Zulu API left alpha the Java versions that openhabian required
were no longer being distributed. As a result this commit changes the
default Java provider on openhabian to be AdoptOpenJDK.

I have personally tested this change and in my use it had performed
without any issues however, that many not be the case for everywhere.

Closes #766

Signed-off-by: Ethan Dye <mrtops03@gmail.com>