Use Zulu API v1 when upgrading Java
Additionally, to clarify what Java version this will install, when
available, the API will prefer downloading a JRE. However, when a JRE is
not available it will use a JDK instead. However, when JDK's are
installed only the essential binaries are linked to update-alternatives
so extra JDK related software is unlinked (i.e. javac).

At the moment, Azul appears to only provide JDK builds for ARM platforms
and so, as such, ARM platforms are most often the ones to not get a JRE
installed and the resulting installation tends to be a bit larger (~100
MB).

One minor inconvenience is that with the current Azul API it does not
allow explicitly choosing builds without the JavaFX additional binaries.
So at times, the Java version installed will have extra JavaFX binaries.
This is usually seen when the architecture being installed is x86 based.

Fixes #714
Fixes #723
See Also #629

Signed-off-by: Ethan Dye <mrtops03@gmail.com>