Refuse to install 64-bit JDK on 32-bit OS
armv8 builds in theory should work on the RPi3+/RPi4, however the
current Raspbian kernel uses a armv7l base which works because armv8 is
backwards compatible. However, this has the effect of making the builds
targeting aarch64 provided by Azul unable to work on the RPi3+/RPi4. On
the other hand, this is not a huge issue as the use of 64-bit Java on
low-memory boards brings few benefits and mostly issues.

Additionally, this adds checking for other platforms to ensure that the
OS is actually 64-bit before trying to install a 64-bit JDK. If the OS
is not actually 64-bit the installer defaults to installing a 32-bit
JDK.

Signed-off-by: Ethan Dye <mrtops03@gmail.com>