Begin Standardization and Rewrite (#938)
* Begin Standardization and Rewrite

Begins a rewrite of the openHABian core code to make it overall safer
and more verbose in failure. Additionally this helps to standardize the
coding format.

This contains major a major rewrite of 'packages.bash',
'nodejs-apps.bash', and other minor fixes in other files as well. This
also contains the proper testing functions using BATS for the rewritten
code.

Fixes #936
Related #937 #949

Signed-off-by: Ethan Dye <mrtops03@gmail.com>

* Documentation and coding guidelines update

Also this includes further changes in terms of the Standardization and
Rewrite.

Fixes #949
Related #519 #625

Signed-off-by: Ethan Dye <mrtops03@gmail.com>