Add build image function to GitHub Actions
The 'Build Image' action was primarily written by Holger Friedrich and
then edited for better functionality in the openHABian core repo.

This also contains other changes to the CI setup, such as moving the
ShellCheck checks out of Travis and instead enforcing them in GitHub
Actions.

Along with theses changes, some improvements to the build script and
install output were also made. These changes to the build script include
using cURL over wget for downloading needed files and also cloning the
openHABian repo in the build process rather than the install process.
This change was made because the repo must be present for offline
install to work, and as such it makes sense to provide it with the image
to reduce points of failure on install.

Fixes #991

Co-authored-by: Holger Friedrich <mail@holger-friedrich.de>
Co-authored-by: Markus Storm <markus.storm@gmx.net>
Signed-off-by: Ethan Dye <mrtops03@gmail.com>