Add automatic release drafting
This primarily solves the issue of having to download and reupload the
image files for releases.

When used, this will let the maintainers simply push a tag in the format
of 'vx.x.x' and the action will run and generate a release draft with
the proper image files alreay attached. All the maintainer needs to do
is wait for it to generate and while waiting make some nice and pretty
release notes to go along with it.

Signed-off-by: Ethan Dye <mrtops03@gmail.com>