Improve CI-Pipeline and other related changes (#1617)
This primarily removes the need for the systemctl replacement in docker
by fully utilizing the initsystem provided by Balena in their docker
images. This system does not work for BATS tests in GitHub Actions
currently. However, the initsystem does work for BATS on my machine in
Docker but I cannot get it to work in the online GitHub Actions
environment.

This also includes a bunch of bug fixes that were found along the way.

Signed-off-by: Ethan Dye <mrtops03@gmail.com>