Change Java providers and add support for Java 17 (#1632)
* Add Java 17 support to API

* Fix ShellCheck error

* Add Java 17 to menu options

Additionally remove AdoptOpenJDK 17 support from API until further
notice. AdoptOpenJDK is currently undergoing a rebranding to Adoptium
and as such they are not fully transitioned and new releases are no
longer being added to the AdoptOpenJDK APT repository. As such until all
releases have been migrated to a new APT repository we will not make any
attempts to support Adoptium releases.

* Add Java 17 to installation tests

* Remove Azul Java 17 on arm32

Azul does not currently have a build of Java 17 for arm32 available and
as such the usage of the arm32 API request for Java 17 will always fail.
Hopefully builds will become available in the future, and because of
this support has been left in the API code.

* Use Zulu APT repo for arm64 Java 17

* Fix package names in `java_zulu_enterprise_apt`

* Fix deprecated Zulu API key

* Install Java using builtin repo

* Use builtin APT repo for JRE

* Fix Java install calls

* Update build.bash for Java install change

* Fix ShellCheck error

* Add Java 17 support

* Add amd64 Java 17 installation test

* Fix error in java-jre.bash

* Cleanup of legacy references

* Update NEWS.md and CHANGELOG.md

* Add Zulu back as insurance until we are confident

* Cleanup some legacy code from zulu functions

* Update NEWS and CHANGELOG dates

Signed-off-by: Ethan Dye <mrtops03@gmail.com>